+++ pages/api/preguntas/[id].ts
@@ -11,7 +11,7 @@ type Data = {
 export default function (req: NextApiRequest, res: NextApiResponse<Data>) {
   switch (req.method) {
     case 'PUT':
-      return updatePositionPregunta(req, res);
+      return updatePregunta(req, res);
 
     default:
       res.status(405).json({ message: 'Method not allowed' });
@@ -19,13 +19,16 @@ export default function (req: NextApiRequest, res: NextApiResponse<Data>) {
   }
 }
 
-const updatePositionPregunta = async (req: NextApiRequest, res: NextApiResponse<Data>) => {
+const updatePregunta = async (req: NextApiRequest, res: NextApiResponse<Data>) => {
   try {
-    const { id, posicion } = req.body;
+    const { id } = req.query;
+    const data = req.body;
+
     const pregunta = await prisma.pregunta.update({
-      where: { id },
-      data: { posicion },
+      where: { id: Number(id) },
+      data,
     });
+
     return res.status(200).json({
       message: 'ok',
       data: pregunta,
@@ -37,3 +40,22 @@ const updatePositionPregunta = async (req: NextApiRequest, res: NextApiResponse<
     });
   }
 };
+
+// const updatePositionPregunta = async (req: NextApiRequest, res: NextApiResponse<Data>) => {
+//   try {
+//     const { id, posicion } = req.body;
+//     const pregunta = await prisma.pregunta.update({
+//       where: { id },
+//       data: { posicion },
+//     });
+//     return res.status(200).json({
+//       message: 'ok',
+//       data: pregunta,
+//     });
+//   } catch (error) {
+//     return res.status(500).json({
+//       message: 'error',
+//       data: error,
+//     });
+//   }
+// };
